Applied patches from Matt Leininger to fix computation of offsets for very large runs (a 32 vs 64 bit issue).
--HG--
extra : convert_revision : svn%3A0fb12191-db4b-0410-8866-e78deaee71e5/trunk%407449

+++ mpqc/src/lib/chemistry/qc/mbpt/csgrad.cc
@@ -566,7 +566,7 @@ MBPT2::compute_cs_grad()
       abort();
     }
 
-  mem->set_localsize(nijmax*nbasis*nbasis*sizeof(double));
+  mem->set_localsize(size_t(nijmax)*nbasis*nbasis*sizeof(double));
   ExEnv::out0() << indent
        << "Size of global distributed array:       "
        << mem->totalsize()
